requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
fidel==1.1.0
langdetect==1.0.9
deep-translator==1.11.4
fastapi==0.104.1
uvicorn==0.24.0